requirements.txt
opencv-python>=4.8.0
scikit-image>=0.20.0
numpy>=1.21.0
streamlit>=1.25.0
Pillow>=9.0.0 